+++ src/components/sbb-link/sbb-link.stories.js
@@ -20,8 +20,6 @@ const wrapperStyle = (context) => {
   return `background-color: ${SbbColorCharcoalDefault};`;
 };
 
-const Template = (args) => <sbb-link {...args}>{args.text}</sbb-link>;
-
 const paragraphStyle = (context) => {
   let color;
 
@@ -33,6 +31,15 @@ const paragraphStyle = (context) => {
 
   return `${color} font-family: ${SbbTypoTypeFaceSbbRoman}; font-weight: normal; line-height: ${SbbTypoLineHeightBodyText}; letter-spacing: ${SbbTypoLetterSpacingBodyText}; font-size: ${SbbTypoScaleDefault}px`;
 };
+
+const Template = (args) => <sbb-link {...args}>{args.text}</sbb-link>;
+
+const IconSlotTemplate = (args) => (
+  <sbb-link {...args}>
+    {args.text}
+    <sbb-icon slot="icon" name={args['icon-name']}></sbb-icon>
+  </sbb-link>
+);
 
 const InlineTemplate = (args, context) => (
   <p style={paragraphStyle(context)}>
@@ -100,6 +107,24 @@ const href = {
 const download = {
   control: {
     type: 'boolean',
+  },
+  table: {
+    category: 'Link',
+  },
+};
+
+const target = {
+  control: {
+    type: 'text',
+  },
+  table: {
+    category: 'Link',
+  },
+};
+
+const rel = {
+  control: {
+    type: 'text',
   },
   table: {
     category: 'Link',
@@ -176,6 +201,15 @@ const eventId = {
   },
 };
 
+const value = {
+  control: {
+    type: 'text',
+  },
+  table: {
+    category: 'Button',
+  },
+};
+
 const defaultArgTypes = {
   text,
   variant,
@@ -183,8 +217,10 @@ const defaultArgTypes = {
   'text-size': textSize,
   'icon-name': iconName,
   'icon-placement': iconPlacement,
-  href: href,
+  href,
   download,
+  target,
+  rel,
   'id-value': idValue,
   'accessibility-label': accessibilityLabel,
   'accessibility-describedby': accessibilityDescribedby,
@@ -194,6 +230,7 @@ const defaultArgTypes = {
   form,
   disabled,
   'event-id': eventId,
+  value,
 };
 
 const defaultArgs = {
@@ -201,19 +238,22 @@ const defaultArgs = {
   variant: variant.options[0],
   negative: false,
   'text-size': textSize.options[1],
-  'icon-name': '',
+  'icon-name': undefined,
   'icon-placement': iconPlacement.options[0],
   href: 'https://github.com/lyne-design-system/lyne-components',
   download: false,
-  'id-value': '',
+  target: undefined,
+  rel: undefined,
+  'id-value': undefined,
   'accessibility-label': 'Travelcards & tickets',
-  'accessibility-describedby': '',
-  'accessibility-labelledby': '',
+  'accessibility-describedby': undefined,
+  'accessibility-labelledby': undefined,
   name: 'Button name',
   type: type.options[0],
-  form: '',
+  form: undefined,
   disabled: false,
   'event-id': 'Event ID for button click',
+  value: undefined,
 };
 
 /* ************************************************* */
@@ -308,7 +348,7 @@ export const BlockButton = Template.bind({});
 BlockButton.argTypes = defaultArgTypes;
 BlockButton.args = {
   ...defaultArgs,
-  href: '',
+  href: undefined,
   'icon-name': 'chevron-small-right-small',
   'icon-placement': iconPlacement.options[1],
 };
@@ -321,12 +361,36 @@ BlockButtonNegative.argTypes = defaultArgTypes;
 BlockButtonNegative.args = {
   ...defaultArgs,
   negative: true,
-  href: '',
+  href: undefined,
   'icon-name': 'chevron-small-right-small',
   'icon-placement': iconPlacement.options[1],
 };
 BlockButtonNegative.documentation = {
   title: 'Block Button Negative',
+};
+
+export const BlockWithSlottedIcon = IconSlotTemplate.bind({});
+BlockWithSlottedIcon.argTypes = defaultArgTypes;
+BlockWithSlottedIcon.args = {
+  ...defaultArgs,
+  'icon-name': 'chevron-small-right-small',
+  'icon-placement': iconPlacement.options[1],
+};
+BlockWithSlottedIcon.documentation = {
+  title: 'Block with slotted icon',
+};
+
+export const BlockLinkOpensInNewWindow = IconSlotTemplate.bind({});
+BlockLinkOpensInNewWindow.argTypes = defaultArgTypes;
+BlockLinkOpensInNewWindow.args = {
+  ...defaultArgs,
+  'icon-name': 'chevron-small-right-small',
+  'icon-placement': iconPlacement.options[1],
+  target: '_blank',
+  'accessibility-label': undefined,
+};
+BlockLinkOpensInNewWindow.documentation = {
+  title: 'Block link opened in new window',
 };
 
 export const Inline = InlineTemplate.bind({});
@@ -358,7 +422,7 @@ InlineButton.args = {
   ...defaultArgs,
   text: 'Show more',
   variant: 'inline',
-  href: '',
+  href: undefined,
 };
 InlineNegative.documentation = {
   title: 'Inline Button',